% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanJobPostings.R
\name{CleanerRunner}
\alias{CleanerRunner}
\title{Cleans the mthml files inside the specified /data/ folder}
\usage{
CleanerRunner(dataFolder = "data/indeed/", outFolder = "data/cleaned_indeed/")
}
\arguments{
\item{dataFolder}{The filepath holding all of your mhtml files}

\item{outFolder}{The filepath to produce the output, default is the same folder}
}
\description{
Cleans the mthml files inside the specified /data/ folder
}
\examples{
\dontrun{
folder = "data/linkedin/"
optionalFolder = "data/cleaned_linkedin"

CleanerRunner(folder, optionalFolder)

}
}
\author{
Anthogonyst
}
