% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanJobPostings.R
\name{CleanMHtml}
\alias{CleanMHtml}
\title{Cleans a mthml file and produces a minified html and json file in the same folder, unless optionally specified}
\usage{
CleanMHtml(fp, out = gsub("\\\\..*$", "", fp))
}
\arguments{
\item{fp}{The filepath to a mhtml file}

\item{out}{The filepath to produce the output, default is the same folder}
}
\description{
Cleans a mthml file and produces a minified html and json file in the same folder, unless optionally specified
}
\examples{
\dontrun{
file = "data/indeed/1Data Scientist Jobs, Employment in Austin, TX _ Indeed.com.mhtml"
optionalFolder = "data/cleaned_file"

CleanMHtml(file, optionalFolder)

file.exists("data/cleaned_file-mini.html")
file.exists("data/cleaned_file-mini.json")

}
}
\author{
Anthogonyst
}
