% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanJobPostings.R
\name{.DetectLineEndings}
\alias{.DetectLineEndings}
\title{The mhtml file splits the file by line length and appends "=" where it exceeds a certain length}
\usage{
.DetectLineEndings(x)
}
\arguments{
\item{x}{A chr vector}
}
\description{
The mhtml file splits the file by line length and appends "=" where it exceeds a certain length
}
\examples{
\dontrun{
before = "some mhtml stuff="
after = "some mhtml stuff"

.DetectLineEndings(before) == after

}
}
\author{
Anthogonyst
}
