% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanJobPostings.R
\name{.FixMHtml}
\alias{.FixMHtml}
\title{Vectorizes the cleaning function and gets rid of the detected line endings}
\usage{
.FixMHtml(x)
}
\arguments{
\item{x}{A list of chr vector}
}
\description{
Vectorizes the cleaning function and gets rid of the detected line endings
}
\examples{
\dontrun{
before = list("first=", "second=", "third=")
after = list("first", "second", "third")

all(mapply(.FixMHtml(before), after, FUN = `==`))

}
}
\author{
Anthogonyst
}
