% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKeywordsDatasets.R
\name{GenerateKeywords}
\alias{GenerateKeywords}
\title{Generates the keywords frequency list by examining nouns and such}
\usage{
GenerateKeywords(
  jobs,
  jobIds,
  writeCsv = NA,
  webster = dictionary,
  captures = captureGroups,
  FUN = GrabLinkedin
)
}
\arguments{
\item{jobs}{The filepath to a file, as parsed by @{FUN}}

\item{jobIds}{The name or id of the file in question}

\item{writeCsv}{Optionally write a csv file by supplying a filepath}

\item{webster}{A dictionary that contains the word bank with columns "pos" and "word"}

\item{captures}{The capture groups from @{webster} in the "pos" column}

\item{FUN}{Data grabbing function to be defined externally; can simply be read.csv}
}
\description{
Generates the keywords frequency list by examining nouns and such
}
\examples{
\dontrun{
sentence = "The big red dog"
Read = function(x) { x }

GenerateKeywords(sentence, "clifford", "data/big-red-dog/", Read)

}
}
\author{
Anthogonyst
}
