% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKeywordsDatasets.R
\name{GrabLinkedin}
\alias{GrabLinkedin}
\title{Defines how to look the Linkedin json data and convert to a vector of all capital letters}
\usage{
GrabLinkedin(x)
}
\arguments{
\item{x}{Filepath}
}
\description{
Defines how to look the Linkedin json data and convert to a vector of all capital letters
}
\author{
Anthogonyst
}
