% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKeywordsDatasets.R
\name{LoadKeywordDatabase}
\alias{LoadKeywordDatabase}
\title{Generates the keywords frequency list by examining nouns and such}
\usage{
LoadKeywordDatabase(keywordsFolder = "data/keywords_linkedin/")
}
\arguments{
\item{keywordsFolder}{The folder with keyword frequencies found by @{GenerateKeywords}}
}
\description{
Generates the keywords frequency list by examining nouns and such
}
\examples{
\dontrun{
df = GenerateKeywords(data, "primary_key", "data/discovered_keywords")

LoadKeywordDatabase("data/discovered_keywords")

}
}
\author{
Anthogonyst
}
