% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{ReadFile}
\alias{ReadFile}
\title{This function reads the data files within the data repository.}
\usage{
ReadFile(name)
}
\arguments{
\item{name}{The name of the file from the data/ folder}
}
\description{
This function reads the data files within the data repository.
}
\examples{
\dontrun{

testData = ReadFile("data.csv")
}
}
\author{
Anthogonyst
}
