% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateKeywordsDatasets.R
\name{SumFreq}
\alias{SumFreq}
\title{Generates the keywords frequency list by examining nouns and such}
\usage{
SumFreq(keywordsListOfLists)
}
\arguments{
\item{keywordsListOfLists}{The keyword data from @{GenerateKeywords} or loaded from @{LoadKeywordDatabase}}
}
\description{
Generates the keywords frequency list by examining nouns and such
}
\examples{
\dontrun{
aggregate = SumFreq(LoadKeywordDatabase("data/discovered_keywords"))

}
}
\author{
Anthogonyst
}
